# ARROW-2380: [Python] Streamline conversions