--- a/cpp/src/arrow/python/common.h
+++ b/cpp/src/arrow/python/common.h
@@ -19,7 +19,9 @@
 #define ARROW_PYTHON_COMMON_H
 
 #include <memory>
+#include <sstream>
 #include <string>
+#include <utility>
 
 #include "arrow/python/config.h"
 
@@ -32,6 +34,15 @@ namespace arrow {
 class MemoryPool;
 
 namespace py {
+
+Status CheckPyError(StatusCode code = StatusCode::UnknownError);
+
+Status PassPyError();
+
+// TODO(wesm): We can just let errors pass through. To be explored later
+#define RETURN_IF_PYERROR() RETURN_NOT_OK(CheckPyError());
+
+#define PY_RETURN_IF_ERROR(CODE) RETURN_NOT_OK(CheckPyError(CODE));
 
 class ARROW_EXPORT PyAcquireGIL {
  public:
@@ -70,6 +81,11 @@ class ARROW_EXPORT OwnedRef {
   OwnedRef(OwnedRef&& other) : OwnedRef(other.detach()) {}
   explicit OwnedRef(PyObject* obj) : obj_(obj) {}
 
+  OwnedRef& operator=(OwnedRef&& other) {
+    obj_ = other.detach();
+    return *this;
+  }
+
   ~OwnedRef() { reset(); }
 
   void reset(PyObject* obj) {
@@ -88,6 +104,8 @@ class ARROW_EXPORT OwnedRef {
   PyObject* obj() const { return obj_; }
 
   PyObject** ref() { return &obj_; }
+
+  operator bool() const { return obj_ != NULLPTR; }
 
  private:
   ARROW_DISALLOW_COPY_AND_ASSIGN(OwnedRef);
@@ -110,36 +128,71 @@ class ARROW_EXPORT OwnedRefNoGIL : public OwnedRef {
   }
 };
 
-struct ARROW_EXPORT PyObjectStringify {
-  OwnedRef tmp_obj;
+// A temporary conversion of a Python object to a bytes area.
+struct ARROW_EXPORT PyBytesView {
   const char* bytes;
   Py_ssize_t size;
 
-  explicit PyObjectStringify(PyObject* obj) {
-    PyObject* bytes_obj;
+  PyBytesView() : bytes(nullptr), size(0), ref(nullptr) {}
+
+  // View the given Python object as binary-like, i.e. bytes
+  static Status FromBinary(PyObject* obj, PyBytesView* out) {
+    return FromBinary(obj, out, "a bytes object");
+  }
+
+  // View the given Python object as string-like, i.e. str or (utf8) bytes
+  // XXX make this non-static?
+  static Status FromString(PyObject* obj, PyBytesView* out, bool check_valid = false) {
     if (PyUnicode_Check(obj)) {
-      bytes_obj = PyUnicode_AsUTF8String(obj);
-      tmp_obj.reset(bytes_obj);
-      bytes = PyBytes_AsString(bytes_obj);
-      size = PyBytes_GET_SIZE(bytes_obj);
-    } else if (PyBytes_Check(obj)) {
-      bytes = PyBytes_AsString(obj);
-      size = PyBytes_GET_SIZE(obj);
+#if PY_MAJOR_VERSION >= 3
+      Py_ssize_t size;
+      // The utf-8 representation is cached on the unicode object
+      const char* data = PyUnicode_AsUTF8AndSize(obj, &size);
+      RETURN_IF_PYERROR();
+      *out = PyBytesView(data, size);
+      return Status::OK();
+#else
+      PyObject* converted = PyUnicode_AsUTF8String(obj);
+      RETURN_IF_PYERROR();
+      *out = PyBytesView(PyBytes_AS_STRING(converted), PyBytes_GET_SIZE(converted),
+                         converted);
+      return Status::OK();
+#endif
     } else {
-      bytes = NULLPTR;
-      size = -1;
+      PyBytesView view;
+      RETURN_NOT_OK(FromBinary(obj, &view, "a string or bytes object"));
+      if (check_valid) {
+        // Check the bytes are valid utf-8
+        OwnedRef decoded(PyUnicode_FromStringAndSize(view.bytes, view.size));
+        RETURN_IF_PYERROR();
+      }
+      *out = std::move(view);
+      return Status::OK();
     }
   }
+
+ protected:
+  PyBytesView(const char* b, Py_ssize_t s, PyObject* obj = nullptr)
+      : bytes(b), size(s), ref(obj) {}
+
+  static Status FromBinary(PyObject* obj, PyBytesView* out,
+                           const std::string& expected_msg) {
+    if (PyBytes_Check(obj)) {
+      *out = PyBytesView(PyBytes_AS_STRING(obj), PyBytes_GET_SIZE(obj));
+      return Status::OK();
+    } else if (PyByteArray_Check(obj)) {
+      *out = PyBytesView(PyByteArray_AS_STRING(obj), PyByteArray_GET_SIZE(obj));
+      return Status::OK();
+    } else {
+      std::stringstream ss;
+      ss << "Expected " << expected_msg << ", got a '" << Py_TYPE(obj)->tp_name
+         << "' object";
+      return Status::TypeError(ss.str());
+    }
+  }
+
+  OwnedRef ref;
 };
-
-Status CheckPyError(StatusCode code = StatusCode::UnknownError);
-
-Status PassPyError();
-
-// TODO(wesm): We can just let errors pass through. To be explored later
-#define RETURN_IF_PYERROR() RETURN_NOT_OK(CheckPyError());
-
-#define PY_RETURN_IF_ERROR(CODE) RETURN_NOT_OK(CheckPyError(CODE));
 
 // Return the common PyArrow memory pool
 ARROW_EXPORT void set_default_memory_pool(MemoryPool* pool);