… cpp/src/arrow/python/CMakeLists.txt
… cpp/src/arrow/python/decimal.h
… cpp/src/arrow/python/helpers.h
… cpp/src/arrow/python/python-test.cc
… cpp/src/arrow/python/python_to_arrow.cc
… python/pyarrow/lib.pyx
… python/pyarrow/tests/test_convert_builtin.py
… python/pyarrow/tests/test_convert_pandas.py
… python/pyarrow/tests/test_feather.py