* Correctly check for overflow when constructing string / binary objects
* Accept bytearrays in the same places as bytes objects
* Other refactorings