@pitrou 2018-04-04
Note to self: I should run the benchmarks when this PR is ready