@pitrou 2018-04-05
Running the benchmarks led me to identify a regression due to instantiating a `std::string` in critical places (fixed).