@BryanCutler 2018-04-05
Thanks @pitrou , looks good!